test: add e2e test user page (#146)
* test: add test for user profile

* chore: add data-testid

* chore: update config

* fix: fix test

* chore: add data-testid attr

* test: add e2e test for user skills

* chore: add data-testid attr

* fix: change localization

* test: add e2e test for user languages

* feat: replace hardcore id to id from cookies

* chore: add data-testid attr, add user id

* test: add e2e user cvs test

---------

+++ src/features/cvs/hooks/use-delete-cv.ts
@@ -5,7 +5,7 @@ import { usePermissions } from "@/hooks/use-permissions"
 import { Cv } from "@/types/graphql-types"
 import { removeRefById } from "@/utils/cache"
 
-export function useDeleteCv(cv: Cv) {
+export function useDeleteCv(cv: Cv, userId?: string) {
   const [deleteCv] = useMutation(DELETE_CV, {
     update(cache) {
       cache.modify({
@@ -37,7 +37,7 @@ export function useDeleteCv(cv: Cv) {
   const { canDeleteCv } = usePermissions()
 
   const handleDelete = async () => {
-    if (!canDeleteCv(cv.user?.id)) return
+    if (!canDeleteCv(cv.user?.id || userId)) return
     await deleteCv({ variables: { cv: { cvId: cv.id } } })
   }
